%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiSourceCopyNumberNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAllNames.MultiSourceCopyNumberNormalization}
\alias{getAllNames.MultiSourceCopyNumberNormalization}
\alias{MultiSourceCopyNumberNormalization.getAllNames}
\alias{getAllNames.MultiSourceCopyNumberNormalization}
\alias{getAllNames,MultiSourceCopyNumberNormalization-method}


\title{Gets the names of all unique samples across all sources}

\description{
 Gets the names of all unique samples across all sources.
}

\usage{\method{getAllNames}{MultiSourceCopyNumberNormalization}(this, ...)}

\arguments{
  \item{...}{Passed to \code{getNames(...)} of each data set.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiSourceCopyNumberNormalization}}.
}
\keyword{internal}
\keyword{methods}
