%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiSourceCopyNumberNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFitAromaUgpFile.MultiSourceCopyNumberNormalization}
\alias{getFitAromaUgpFile.MultiSourceCopyNumberNormalization}
\alias{MultiSourceCopyNumberNormalization.getFitAromaUgpFile}
\alias{getFitAromaUgpFile.MultiSourceCopyNumberNormalization}
\alias{getFitAromaUgpFile,MultiSourceCopyNumberNormalization-method}


\title{Gets the UGP file specifying the common set of loci to normalize at}

\description{
 Gets the UGP file specifying the common set of loci to normalize at.
}

\usage{\method{getFitAromaUgpFile}{MultiSourceCopyNumberNormalization}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[aroma.core]{AromaUgpFile}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiSourceCopyNumberNormalization}}.
}
\keyword{internal}
\keyword{methods}
