%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiSourceCopyNumberNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInputDataSets.MultiSourceCopyNumberNormalization}
\alias{getInputDataSets.MultiSourceCopyNumberNormalization}
\alias{MultiSourceCopyNumberNormalization.getInputDataSets}
\alias{getInputDataSets.MultiSourceCopyNumberNormalization}
\alias{getInputDataSets,MultiSourceCopyNumberNormalization-method}


\title{Gets the list of data sets to be normalized}

\description{
 Gets the list of data sets to be normalized.
}

\usage{\method{getInputDataSets}{MultiSourceCopyNumberNormalization}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiSourceCopyNumberNormalization}}.
}
\keyword{internal}
\keyword{methods}
