%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultiSourceCopyNumberNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSubsetToFit.MultiSourceCopyNumberNormalization}
\alias{getSubsetToFit.MultiSourceCopyNumberNormalization}
\alias{MultiSourceCopyNumberNormalization.getSubsetToFit}
\alias{getSubsetToFit.MultiSourceCopyNumberNormalization}
\alias{getSubsetToFit,MultiSourceCopyNumberNormalization-method}


\title{Gets subset of (smoothing) units for fitting the model}

\description{
 Gets subset of (smoothing) units for fitting the model.
}

\usage{\method{getSubsetToFit}{MultiSourceCopyNumberNormalization}(this, ..., verbose=FALSE)}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of unit indices.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultiSourceCopyNumberNormalization}}.
}
\keyword{internal}
\keyword{methods}
