%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{allocate.AromaSignalBinaryFile}
\alias{as.character.AromaSignalBinaryFile}
\alias{byName.AromaSignalBinarySet}
\alias{findByName.AromaSignalBinarySet}
\alias{fromFile.AromaSignalBinaryFile}
\alias{getChipType.AromaSignalBinaryFile}
\alias{getChipType.AromaSignalBinarySet}
\alias{getFilenameExtension.AromaSignalBinaryFile}
\alias{getPlatform.AromaSignalBinaryFile}
\alias{getPlatform.AromaSignalBinarySet}
\alias{nbrOfUnits.AromaSignalBinaryFile}
\alias{validate.AromaSignalBinarySet}
\alias{byName.AromaFreqBCnBinarySet}
\alias{byName.AromaTotalCnBinarySet}
\alias{getAromaUgpFile.AromaSignalBinaryFile}
\alias{getAromaUgpFile.AromaSignalBinarySet}
\alias{getUnitNamesFile.AromaSignalBinaryFile}
\alias{extractRawCopyNumbers.AromaTotalCnBinaryFile}
\alias{getAsteriskTags.TotalCnSmoothing}
\alias{getOutputFiles.TotalCnSmoothing}
\alias{getParameters.TotalCnSmoothing}
\alias{getPath.TotalCnSmoothing}
\alias{getRootPath.TotalCnSmoothing}
\alias{getTargetPositions}
\alias{getTargetPositions.TotalCnSmoothing}
\alias{getTargetUgpFile}
\alias{getTargetUgpFile.TotalCnSmoothing}
\alias{process.TotalCnSmoothing}
\alias{TotalCnSmoothing}
\alias{fitPrincipalCurve}
\alias{fitPrincipalCurve.data.frame}
\alias{fitPrincipalCurve.matrix}
\alias{get.lam.hb}
\alias{makeSmoothSplinePredict}
\alias{makeSmoothSplinePredict.numeric}
\alias{normalizePrincipalCurve}
\alias{normalizePrincipalCurve.data.frame}
\alias{normalizePrincipalCurve.matrix}
\alias{principal.curve.hb}
\alias{as.character.MultiSourceCopyNumberNormalization}
\alias{extractTupleOfDataFiles}
\alias{fitOne}
\alias{getAllNames}
\alias{getFitAromaUgpFile}
\alias{getInputDataSets}
\alias{getOutputDataSets}
\alias{getOutputDataSets.MultiSourceCopyNumberNormalization}
\alias{getOutputPaths}
\alias{getOutputPaths.MultiSourceCopyNumberNormalization}
\alias{getParameters.MultiSourceCopyNumberNormalization}
\alias{getParametersAsString.MultiSourceCopyNumberNormalization}
\alias{getSmoothedDataSets}
\alias{getSubsetToFit}
\alias{nbrOfDataSets}
\alias{nbrOfDataSets.MultiSourceCopyNumberNormalization}
\alias{normalizeOne}
\alias{normalizeOne.MultiSourceCopyNumberNormalization}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
