%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  XYCurveNormalization.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{XYCurveNormalization}
\docType{class}
\alias{XYCurveNormalization}


 \title{The XYCurveNormalization class}

 \description{
  Package:  aroma.cn \cr
\bold{Class XYCurveNormalization}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[aroma.cn]{AbstractCurveNormalization}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{XYCurveNormalization}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{XYCurveNormalization}\cr
extends \emph{\link[aroma.cn]{AbstractCurveNormalization}}\cr


 }

 \usage{XYCurveNormalization(...)}

 \arguments{
  \item{...}{Arguments passed to \code{\link{AbstractCurveNormalization}}.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\emph{No methods defined}.


 \bold{Methods inherited from AbstractCurveNormalization}:\cr
backtransformOne, fitOne, getDataSets, getFullName, getInputDataSet, getName, getOutputDataSet, getPairedDataSet, getPath, getRootPath, getSubsetToFit, getTags, getTargetDataSet, nbrOfFiles, process, setTags

 \bold{Methods inherited from Object}:\cr
asThis, getChecksum, $, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Henrik Bengtsson}
\keyword{classes}
