%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.cn-package}
\alias{aroma.cn-package}
\alias{aroma.cn}
\docType{package}

\title{Package aroma.cn}


 \description{
   Package for analysis of copy-number estimates obtained from various platforms.

   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires the \pkg{aroma.core} package.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{http://www.aroma-project.org/}.
 }

 \section{To get started}{
   To get started, see:
   \enumerate{
     \item ...
   }
 }

 \author{
   Henrik Bengtsson.
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.

   The development code of the packages is under a private licence
   (where applicable) and patches sent to the author fall under the
   latter license, but will be, if incorporated, released under the
   "release" license above.
 }

 \references{
 
Please cite aroma.cn as

  H. Bengtsson, A. Ray, P. Spellman and T.P. Speed. A single-sample
  method for normalizing and combining full-resolutioncopy numbers from
  multiple platforms, labs and analysis methods, Bioinformatics, 2009

  H. Bengtsson; K. Simpson; J. Bullard; K. Hansen. aroma.affymetrix: A
  generic framework in R for analyzing small to very large Affymetrix
  data sets in bounded memory, Tech Report 745, Department of
  Statistics, University of California, Berkeley, February 2008

  H. Bengtsson, R. Irizarry, B. Carvalho, & T.P. Speed. Estimation and
  assessment of raw copy numbers at the single locus level,
  Bioinformatics, 2008


 }
\keyword{package}
